Fix compilation errors in DatabaseService and EncryptionService
- Change Convert.ToInt64 to Convert.ToInt32 in DatabaseService for correct return type
- Replace null parameters with new StringBuilder() in DPAPI functions to fix null reference warnings
- Ensure all null warnings are resolved while maintaining functionality

+++ PasswordManager/Services/DatabaseService.cs
@@ -84,7 +84,7 @@ namespace PasswordManager.Services
             command.Parameters.AddWithValue("@LastReminderSent", entry.LastReminderSent == DateTime.MinValue ? default(string) : entry.LastReminderSent.ToString("o"));
 
             var result = command.ExecuteScalar();
-            return Convert.ToInt64(result);
+            return Convert.ToInt32(result);
         }
 
         public bool UpdatePasswordEntry(PasswordEntry entry)
